% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balls_overs.R
\name{balls_to_overs}
\alias{balls_to_overs}
\title{Convert Balls to Overs (Six Ball)}
\usage{
balls_to_overs(balls)
}
\arguments{
\item{balls}{number of balls bowled/faced.}
}
\value{
number of six ball overs this equates too.
}
\description{
Convert numbers of balls as it equates in terms of six ball overs.
}
\examples{
balls_to_overs(balls = 6)
balls_to_overs(balls = 17)

}
