% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat_raw_df}
\alias{bat_raw_df}
\title{Batters Dataset}
\format{
A data frame with 3 rows and 5 variables:
\describe{
\item{Player}{Name of Player}
\item{Inns}{Numbers of Innings undertaken by Player}
\item{NO}{Numbers of Not Outs by Player}
\item{Runs_Scored}{Numbers of Runs Scored by Player}
\item{Balls_Faced}{Numbers of Balls Faced by Player}
}
}
\usage{
bat_raw_df
}
\description{
A dataset containing basic data about batters
}
\keyword{datasets}
