% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bowl_raw_df}
\alias{bowl_raw_df}
\title{Bowling Dataset}
\format{
A data frame with 3 rows and 4 variables:
\describe{
\item{Player}{Name of Player}
\item{Balls_Bowled}{Numbers of Balls Bowled by Player}
\item{Runs_Conceded}{Numbers of Runs Conceded by Player}
\item{Wickets}{Numbers of Wickets taken by Player}
}
}
\usage{
bowl_raw_df
}
\description{
A dataset containing basic data about bowlers
}
\keyword{datasets}
