% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling_basics.R
\name{bowl_sr}
\alias{bowl_sr}
\title{Bowler Strike Rate}
\usage{
bowl_sr(balls_bowled, wickets_taken)
}
\arguments{
\item{balls_bowled}{number of balls bowled. Data in terms of six ball overs
please convert to \code{\link{overs_to_balls}} to get it terms of \verb{balls bowled}}

\item{wickets_taken}{total wickets taken across the overs bowled}
}
\value{
Number of balls per wicket taken across overs bowled
}
\description{
Calculates Bowlers' number of balls per wicket taken across overs bowled
}
\examples{
bowl_sr(balls_bowled = 3830, wickets_taken = 112)
bowl_sr(balls_bowled = overs_to_balls(overs = 1651.2), wickets_taken = 243)
}
