% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricket.functions-package.R
\docType{package}
\name{cricket.functions-package}
\alias{cricket.functions}
\alias{cricket.functions-package}
\title{cricket.functions: Useful Functions for Cricket Analysis}
\description{
Useful Functions for Cricket Analysis. This package will help you do calculations specific to cricket in a tidy & simple manner.
}
\author{
\strong{Maintainer}: Luke Lockley \email{luke.lockley@btinternet.com} (\href{https://orcid.org/0000-0002-7028-1499}{ORCID})

}
\keyword{internal}
