% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howzatR-package.R
\docType{package}
\name{howzatR-package}
\alias{howzatR}
\alias{howzatR-package}
\title{howzatR: Useful Functions for Cricket Analysis}
\description{
Helping to calculate cricket specific problems in a tidy & simple manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lukelockley/howzatR}
  \item Report bugs at \url{https://github.com/lukelockley/howzatR/issues}
}

}
\author{
\strong{Maintainer}: Luke Lockley \email{luke.lockley@btinternet.com} (\href{https://orcid.org/0000-0002-7028-1499}{ORCID})

}
\keyword{internal}
