% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balls_overs.R
\name{overs_to_balls}
\alias{overs_to_balls}
\title{Convert Overs (Six Ball) to Balls}
\usage{
overs_to_balls(overs)
}
\arguments{
\item{overs}{number of six ball overs bowled/faced.}
}
\value{
number of six ball overs this equates too.
}
\description{
Convert Overs (Six Ball) to Balls
}
\examples{
overs_to_balls(overs = 8.2)
overs_to_balls(overs = 10)

}
